refactor: consolidate into single entry point (closes #30)
- Replace index.ts stub with full generation logic from generate-real-pdfs.ts
- Write PDFs to /tmp instead of output/, delete after generation
- Add TODO (#5) comment for print-before-delete cycle
- Update generate script to target dist/index.js
- Delete src/generate-real-pdfs.ts
